refactor(frontend): extract org field components from organizations detail page (516 -> 325 lines)
